Rust based runtime, forked from cortex-m-rt
Sets up the basis for a rust based runtime forked from the cortex-m-rt
project. This allows loading elfs directly with a debug probe and
running them on evk boards.

A next step would be to support running from flash with the boot config
table setup and some things copied around as done in the imxrt1062-rt.

This requires using the load memory addresses as well as address symbols
defined in a flash or hyperflash specific linker script along with some
extra copy operations in the Reset handler. I believe this is all doable
in Rust still, but time will tell.