Use an absolute address for __dcd
LLVM's lld and GNU's ld have different ways of handling assignments in
output sections. Unless we specify ABSOLUTE, ld treats the number '0' as
a relative address from the section start, 0x6000_0000. On the other
hand, lld treats '0' as if it were written with ABSOLUTE, and it ignores
the ABSOLUTE function. So depending on your linker, __dcd would change
values.

This commit forces an absolute number for __dcd, ensuring a consistent
value no matter the linker.