Ensure __pre_init remains in binary
GNU's LTO has a tendency to remove __pre_init, which is written in
inline assembly. It doesn't realize that the reset handler references
this symbol, because the reset handler is also written in inline
assembly. Not sure why LLVM's linker doesn't also optimize it away, but
this commit ensures that __pre_init remains in the output file.